union owner and permissions to ownership
