+++ packages/api/src/graphql/ownership/mutation/set_ownership.ts
@@ -2,7 +2,7 @@ import pkg from '@apollo/client';
 const { gql } = pkg;
 import { NetworkResult, handleRequest } from "../../../utils/network.js";
 import client from "../../client.js";
-import { Owner } from "../../types/ownership.js";
+import { Ownership } from "../../types/ownership.js";
 
 const SET_OWNER = gql`
   mutation PermissionOwn(
@@ -25,22 +25,15 @@ const SET_OWNER = gql`
   }
 `;
 
-interface OwnershipResponse {
-  userName: string;
-  groupName: string;
-}
-
 export const setOwner = async (
   databaseName: string,
   collectionName: string,
   docId: string | undefined,
   grouping: string,
   newOwner: string
-): Promise<NetworkResult<Owner>> => {
+): Promise<NetworkResult<Ownership>> => {
   return handleRequest(async () => {
-    const { data, errors } = await client.mutate<{
-      permissionSet: OwnershipResponse;
-    }>({
+    const { data, errors } = await client.mutate({
       mutation: SET_OWNER,
       variables: {
         databaseName,
@@ -56,10 +49,7 @@ export const setOwner = async (
     if (response) {
       return {
         type: "success",
-        data: {
-          groupName: response.groupName,
-          userName: response.userName,
-        },
+        data: response.permissionOwn
       };
     } else {
       return {
