+++ packages/api/src/graphql/types/index.ts
@@ -7,8 +7,7 @@ export {
 export { Database, DatabaseStatus } from "./database.js";
 export { DocumentEncoded } from "./document.js";
 export { MerkleWitness } from "./merkle-tree.js";
-export { Owner } from "./ownership.js";
-export { PermissionSet, Permissions } from "./permission.js";
+export { PermissionSet, Permissions, Ownership } from "./ownership.js";
 export { SchemaField, Schema } from "./schema.js";
 export { Session } from "./session.js";
 export { User } from "./user.js";
